+++ instrumentation/log4j/log4j-appender-2.17/library/src/main/java/io/opentelemetry/instrumentation/log4j/appender/v2_17/internal/LogEventMapper.java
@@ -12,15 +12,14 @@ import static io.opentelemetry.semconv.CodeAttributes.CODE_FILE_PATH;
 import static io.opentelemetry.semconv.CodeAttributes.CODE_FUNCTION_NAME;
 import static io.opentelemetry.semconv.CodeAttributes.CODE_LINE_NUMBER;
 import static io.opentelemetry.semconv.OtelAttributes.OTEL_EVENT_NAME;
-import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.toList;
 
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.logs.LogRecordBuilder;
 import io.opentelemetry.api.logs.Severity;
 import io.opentelemetry.context.Context;
+import io.opentelemetry.instrumentation.api.config.IncludeExclude;
 import io.opentelemetry.instrumentation.api.internal.cache.Cache;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.Supplier;
@@ -65,8 +64,7 @@ public final class LogEventMapper<T> {
   private final boolean captureMarkerAttribute;
   private final boolean captureTemplate;
   private final boolean captureArguments;
-  private final List<AttributeKey<String>> captureContextDataAttributeKeys;
-  private final boolean captureAllContextDataAttributes;
+  @Nullable private final IncludeExclude contextDataAttributes;
   private final boolean v3Preview;
 
   @SuppressWarnings("TooManyParameters")
@@ -78,7 +76,7 @@ public final class LogEventMapper<T> {
       boolean captureMarkerAttribute,
       boolean captureTemplate,
       boolean captureArguments,
-      List<String> captureContextDataAttributes,
+      @Nullable IncludeExclude contextDataAttributes,
       boolean v3Preview) {
 
     this.contextDataAccessor = contextDataAccessor;
@@ -88,19 +86,10 @@ public final class LogEventMapper<T> {
     this.captureMarkerAttribute = captureMarkerAttribute;
     this.captureTemplate = captureTemplate;
     this.captureArguments = captureArguments;
-    this.captureAllContextDataAttributes =
-        captureContextDataAttributes.size() == 1 && captureContextDataAttributes.get(0).equals("*");
-    if (captureAllContextDataAttributes) {
-      this.captureContextDataAttributeKeys = emptyList();
-    } else {
-      List<AttributeKey<String>> keys = new ArrayList<>(captureContextDataAttributes.size());
-      for (String key : captureContextDataAttributes) {
-        if (!OTEL_EVENT_NAME.getKey().equals(key)) {
-          keys.add(getContextDataAttributeKey(key));
-        }
-      }
-      this.captureContextDataAttributeKeys = keys;
-    }
+    this.contextDataAttributes =
+        contextDataAttributes == null || contextDataAttributes.isEmpty()
+            ? null
+            : contextDataAttributes;
     this.v3Preview = v3Preview;
   }
 
@@ -249,21 +238,17 @@ public final class LogEventMapper<T> {
       builder.setEventName(otelEventName);
     }
 
-    if (captureAllContextDataAttributes) {
-      contextDataAccessor.forEach(
-          contextData,
-          (key, value) -> {
-            if (!OTEL_EVENT_NAME.getKey().equals(key)) {
-              builder.setAttribute(getContextDataAttributeKey(key), value);
-            }
-          });
+    if (contextDataAttributes == null) {
       return;
     }
 
-    for (AttributeKey<String> attributeKey : captureContextDataAttributeKeys) {
-      String value = contextDataAccessor.getValue(contextData, attributeKey.getKey());
-      builder.setAttribute(attributeKey, value);
-    }
+    contextDataAccessor.forEach(
+        contextData,
+        (key, value) -> {
+          if (!OTEL_EVENT_NAME.getKey().equals(key) && contextDataAttributes.matches(key)) {
+            builder.setAttribute(getContextDataAttributeKey(key), value);
+          }
+        });
   }
 
   public static AttributeKey<String> getContextDataAttributeKey(String key) {